Bimap-envirAPI {AnnotationDbi}	R Documentation

Environment-like API for Bimap objects

Description

These methods allow the user to manipulate any Bimap object as if it was an environment. This environment-like API is provided for backward compatibility with the traditional environment-based maps.

Usage


  ls(name, pos = -1L, envir = as.environment(pos), all.names = FALSE,
     pattern, sorted = TRUE)
  exists(x, where, envir, frame, mode, inherits)
  get(x, pos, envir, mode, inherits)
  #x[[i]]
  #x$name

  ## Converting to a list
  mget(x, envir, mode, ifnotfound, inherits)
  eapply(env, FUN, ..., all.names, USE.NAMES)
  #contents(object, all.names)

  ## Additional convenience method
  sample(x, size, replace=FALSE, prob=NULL, ...)

Arguments

name
A Bimap object for ls. A key as a literal character string or a name (possibly backtick quoted) for x$name.

pos, all.names, USE.NAMES, where, frame, mode, inherits
Ignored.

envir
Ignored for ls. A Bimap object for mget, get and exists.

pattern
An optional regular expression. Only keys matching 'pattern' are returned.

x
The key(s) to search for for exists, get and mget. A Bimap object for [[ and x$name. A Bimap object or an environment for sample.

i
Single key specifying the map element to extract.

ifnotfound
A value to be used if the key is not found. Only NA is currently supported.

env
A Bimap object.

FUN
The function to be applied (see original eapply for environments for the details).

...
Optional arguments to FUN.

size
Non-negative integer giving the number of map elements to choose.

replace
Should sampling be with replacement?

prob
A vector of probability weights for obtaining the elements of the map being sampled.

sorted
logical(1). When TRUE (default), return primary keys in sorted order.

See Also

ls, exists, get, mget, eapply, contents, sample, BimapFormatting, Bimap

Examples


  library(hgu95av2.db)
  x <- hgu95av2CHRLOC

  ls(x)[1:3]
  exists(ls(x)[1], x)
  exists("titi", x)
  get(ls(x)[1], x)
  x[[ls(x)[1]]]
  x$titi # NULL

  mget(ls(x)[1:3], x)
  eapply(x, length)
  contents(x)

  sample(x, 3)

[Package AnnotationDbi version 1.44.0 Index]