GOID {AnnotationDbi}	R Documentation

Descriptions of available values for columns and keytypes for GO.db.

Description

This manual page enumerates the kinds of data represented by the values returned when the user calls columns or keytypes

Details

All the possible values for columns and keytypes are listed below.

GOID:

GO Identifiers

DEFINITION:

The definition of a GO Term

ONTOLOGY:

Which of the three Gene Ontologies (BP, CC, or MF)

TERM:

The actual GO term

To get the latest information about the date stamps and source URLS for the data used to make an annotation package, please use the metadata method as shown in the example below.

Author(s)

Marc Carlson

Examples


  library(GO.db)
  ## List the possible values for columns
  columns(GO.db)
  ## List the possible values for keytypes
  keytypes(GO.db)
  ## get some values back
  keys <- head(keys(GO.db))
  keys
  select(GO.db, keys=keys, columns=c("TERM","ONTOLOGY"),
  keytype="GOID")

  ## More infomation about the dates and original sources for these data:
  metadata(GO.db)

[Package AnnotationDbi version 1.44.0 Index]