GOTerms-class {AnnotationDbi}	R Documentation

Class "GOTerms"

Description

A class to represent Gene Ontology nodes

Objects from the Class

Objects can be created by calls of the form GOTerms(GOId, term, ontology, definition, synonym, secondary). GOId, term, and ontology are required.

Slots

GOID:

Object of class "character" A character string for the GO id of a primary node.

Term:

Object of class "character" A character string that defines the role of gene product corresponding to the primary GO id.

Ontology:

Object of class "character" Gene Ontology category. Can be MF - molecular function, CC - cellular component, or BP - biological process.

Definition:

Object of class "character" Further definition of the ontology of the primary GO id.

Synonym:

Object of class "character" other ontology terms that are considered to be synonymous to the primary term attached to the GO id (e.g. "type I programmed cell death" is a synonym of "apoptosis"). Synonymous here can mean that the synonym is an exact synonym of the primary term, is related to the primary term, is broader than the primary term, is more precise than the primary term, or name is related to the term, but is not exact, broader or narrower.

Secondary:

Object of class "character" GO ids that are secondary to the primary GO id as results of merging GO terms so that One GO id becomes the primary GO id and the rest become the secondary.

Methods

GOID

signature(object = "GOTerms"): The get method for slot GOID.

Term

signature(object = "GOTerms"): The get method for slot Term.

Ontology

signature(object = "GOTerms"): The get method for slot Ontology.

Definition

signature(object = "GOTerms"): The get method for slot Definition.

Synonym

signature(object = "GOTerms"): The get method for slot Synonym.

Secondary

signature(object = "GOTerms"): The get method for slot Secondary.

show

signature(x = "GOTerms"): The method for pretty print.

Note

GOTerms objects are used to represent primary GO nodes in the SQLite-based annotation data package GO.db

References

http://www.geneontology.org/

See Also

makeGOGraph shows how to make GO mappings into graphNEL objects.

Examples


  gonode <- new("GOTerms", GOID="GO:1234567", Term="Test", Ontology="MF",
                          Definition="just for testing")
  GOID(gonode)
  Term(gonode)
  Ontology(gonode)

  ##Or you can just use these methods on a GOTermsAnnDbBimap
## Not run: ##I want to show an ex., but don't want to require GO.db
  require(GO.db)
  FirstTenGOBimap <- GOTERM[1:10] ##grab the 1st ten
  Term(FirstTenGOBimap)

  ##Or you can just use GO IDs directly
  ids = keys(FirstTenGOBimap)
  Term(ids)

## End(Not run)

[Package AnnotationDbi version 1.44.0 Index]