HOMO_SAPIENS {AnnotationDbi}	R Documentation

Descriptions of available values for columns and keytypes for inparanpoid packages.

Description

When the user calls columns or keytypes for an inparanoid package, the columns and keytypes methods will give the full genus and species names of all the organisms that are available.

Details

All the possible values for columns and keytypes are listed below.

ACYRTHOSIPHON_PISUM:

the pea aphid

AEDES_AEGYPTI:

a mosquito that can spread the dengue fever, Chikungunya and yellow fever viruses, and other diseases

ANOPHELES_GAMBIAE:

a mosquito notorious as a vector for malaria

APIS_MELLIFERA:

the western honey bee

ARABIDOPSIS_THALIANA:

the thale cress

ASPERGILLUS_FUMIGATUS:

a fungus that causes disease in immunodeficient individuals

BATRACHOCHYTRIUM_DENDROBATIDIS:

a chytrid fungus that causes the disease chytridiomycosis

BOMBYX_MORI:

the silk worm

BOS_TAURUS:

domestic cattle

BRANCHIOSTOMA_FLORIDAE:

a lancelet (amphioxus)

BRUGIA_MALAYI:

a nematode (roundworm), one of the three causative agents of lymphatic filariasis

CAENORHABDITIS_BRENNERI:

a small nematode, closely related to the model organism Caenorhabditis elegans

CAENORHABDITIS_BRIGGSAE:

a small nematode, closely related to Caenorhabditis elegans

CAENORHABDITIS_ELEGANS:

a small nematode

CAENORHABDITIS_JAPONICA:

a gonochoristic (male-female) species related to C. elegans

CAENORHABDITIS_REMANEI:

a species of nematode (gonochoristic)

CANDIDA_ALBICANS:

a diploid fungus that grows both as yeast and filamentous cells and a causal agent of opportunistic oral and genital infections in humans

CANDIDA_GLABRATA:

a haploid yeast of the genus Candida

CANIS_FAMILIARIS:

domestic dog

CAPITELLA_SPI:

a polychaete worm

CAVIA_PORCELLUS:

Guinea pig

CHLAMYDOMONAS_REINHARDTII:

a single celled green alga

CIONA_INTESTINALIS:

a urochordata (sea squirt), a tunicate widely distributed in Northern European waters

CIONA_SAVIGNYI:

a urochordata (sea squirt)

COCCIDIOIDES_IMMITIS:

a pathogenic fungus that resides in the soil

COPRINOPSIS_CINEREUS:

a species of mushroom

CRYPTOCOCCUS_NEOFORMANS:

an encapsulated yeast that can live in both plants and animals

CRYPTOSPORIDIUM_HOMINIS:

an obligate parasite of humans that can colonize the gastrointestinal tract

CRYPTOSPORIDIUM_PARVUM:

one of several protozoal species that cause cryptosporidiosis, a parasitic disease of the mammalian intestinal tract

CULEX_PIPIENS:

the common house mosquito

CYANIDIOSCHYZON_MEROLAE:

a an algae that is the main organism in red tide

DANIO_RERIO:

the zebrafish

DAPHNIA_PULEX:

the most common species of water flea

DEBARYOMYCES_HANSENII:

a yeast that tolerates high concentrations of salt and is related to yeasts that cause disease, including Candida albicans

DICTYOSTELIUM_DISCOIDEUM:

a species of soil-living amoeba, AKA a slime mold

DROSOPHILA_ANANASSAE:

a fruit fly

DROSOPHILA_GRIMSHAWI:

a fruit fly

DROSOPHILA_MELANOGASTER:

a fruit fly

DROSOPHILA_MOJAVENSIS:

a fruit fly

DROSOPHILA_PSEUDOOBSCURA:

a fruit fly

DROSOPHILA_VIRILIS:

a fruit fly

DROSOPHILA_WILLISTONI:

a fruit fly

ENTAMOEBA_HISTOLYTICA:

an anaerobic parasitic protozoan

EQUUS_CABALLUS:

domestic horse

ESCHERICHIA_COLIK12:

a laboratory strain of coliform bacteria

FUSARIUM_GRAMINEARUM:

a fungus that attacks cereal grains

GALLUS_GALLUS:

domsticated chicken

GASTEROSTEUS_ACULEATUS:

three spined stickleback fish

GIARDIA_LAMBLIA:

a flagellated protozoan parasite

HELOBDELLA_ROBUSTA:

a leech

IXODES_SCAPULARIS:

the black legged deer tick, a vector for lyme disease

KLUYVEROMYCES_LACTIS:

yeast commonly used for genetic studies

LEISHMANIA_MAJOR:

a species of Leishmania, associated with zoonotic cutaneous leishmaniasis

LOTTIA_GIGANTEA:

a species of sea snail, a true limpet, a marine gastropod mollusc

MACACA_MULATTA:

the rhesus Macaque

MAGNAPORTHE_GRISEA:

rice blast fungus

MONODELPHIS_DOMESTICA:

grey short tailed opossum

MONOSIGA_BREVICOLLIS:

a marine choanoflagellate

MUS_MUSCULUS:

lab mouse

NASONIA_VITRIPENNIS:

a small pteromalid parasitoid wasp

NEMATOSTELLA_VECTENSIS:

the starlet sea anemone

NEUROSPORA_CRASSA:

a type of red bread mould

ORNITHORHYNCHUS_ANATINUS:

the platypus

ORYZA_SATIVA:

rice

ORYZIAS_LATIPES:

medaka fish

OSTREOCOCCUS_TAURI:

a unicellular coccoid or spherically shaped green alga

PAN_TROGLODYTES:

chimp

PEDICULUS_HUMANUS:

a species of lice that infects humans

PHYSCOMITRELLA_PATENS:

a moss (Bryophyta) used as a model organism for studies on plant evolution

PHYTOPHTHORA_RAMORUM:

the oomycete plant pathogen (sudden oak death)

PHYTOPHTHORA_SOJAE:

an oomycete and a soil-borne plant pathogen that causes stem and root rot of soybean

PLASMODIUM_FALCIPARUM:

a protozoan parasite that causes malaria

PLASMODIUM_VIVAX:

a protozoal parasite and a human pathogen that causes a more benign malaria

PONGO_PYGMAEUS:

the Bornean orangutan

POPULUS_TRICHOCARPA:

black cottonwood; also known as western balsam poplar or California poplar

PRISTIONCHUS_PACIFICUS:

a diplogastrid nematode

PUCCINIA_GRAMINIS:

stem, black or cereal rusts

RATTUS_NORVEGICUS:

common lab rat

RHIZOPUS_ORYZAE:

a fungus that lives worldwide in dead organic matter. An opportunistic human pathogen

SACCHAROMYCES_CEREVISIAE:

brewers yeast

SCHISTOSOMA_MANSONI:

a significant parasite of humans, a trematode that is one of the major agents of the disease schistosomiasis

SCHIZOSACCHAROMYCES_POMBE:

fission yeast

SCLEROTINIA_SCLEROTIORUM:

an omnivorous fungal plant pathogen

SORGHUM_BICOLOR:

sorghum

STAGONOSPORA_NODORUM:

a fungal leaf spot disease

STRONGYLOCENTROTUS_PURPURATUS:

the purple sea urchin

TAKIFUGU_RUBRIPES:

Japanese pufferfish

TETRAHYMENA_THERMOPHILA:

a single celled cilliate

TETRAODON_NIGROVIRIDIS:

green spotted pufferfish (fresh water)

THALASSIOSIRA_PSEUDONANA:

a species of marine centric diatom

THEILERIA_ANNULATA:

a tickborne protozoan pathogen which is a major cause of livestock disease in sub-tropical regions

THEILERIA_PARVA:

a parasitic protozoan, that causes East Coast fever (theileriosis) in cattle

TRIBOLIUM_CASTANEUM:

the red flour beetle

TRICHOMONAS_VAGINALIS:

an anaerobic, flagellated protozoan

TRICHOPLAX_ADHAERENS:

Trichoplax adhaerens represents the simplest known animal, with the smallest known animal genome

TRYPANOSOMA_CRUZI:

a species of parasitic euglenoid trypanosomes. This species causes the trypanosomiasis diseases in humans and animals in America.

USTILAGO_MAYDIS:

a pathogenic plant fungus that causes smut disease on maize

XENOPUS_TROPICALIS:

Western clawed frog

YARROWIA_LIPOLYTICA:

Yarrowia lipolytica is a "non-conventional" species of yeast, often used in genetic research because it differs from other well-studied species

To get the latest information about the date stamps and source URLS for the data used to make an annotation package, please use the metadata method as shown in the example below.

Author(s)

Marc Carlson

Examples


  library(hom.Hs.inp.db)
  ## List the possible values for columns
  columns(hom.Hs.inp.db)
  ## List the possible values for keytypes
  keytypes(hom.Hs.inp.db)
  ## get some values back
  keys <- head(keys(hom.Hs.inp.db, keytype="HOMO_SAPIENS"))
  keys
  select(hom.Hs.inp.db, keys=keys, columns=c("BOS_TAURUS","EQUUS_CABALLUS"),
  keytype="HOMO_SAPIENS")

  ## More infomation about the dates and original sources for these data:
  metadata(hom.Hs.inp.db)

[Package AnnotationDbi version 1.44.0 Index]