ACCNUM {AnnotationDbi}	R Documentation

Descriptions of available values for columns and keytypes.

Description

This manual page enumerates the kinds of data represented by the values returned when the user calls columns or keytypes

Details

All the possible values for columns and keytypes are listed below. Users will have to actually use these methods to learn which of the following possible values actually apply in their case.

ACCNUM:

GenBank accession numbers

ALIAS:

Commonly used gene symbols

ARACYC:

KEGG Identifiers for arabidopsis as indicated by aracyc

ARACYCENZYME:

Aracyc enzyme names as indicated by aracyc

CHR:

Chromosome (deprecated for Bioc > 3.1) For this information you should look at a TxDb or OrganismDb object and search for an appropriate field like TXCHROM, EXONCHROM or CDSCHROM. This information can also be retrieved from these objects using an appropriate range based accesor like transcripts, transcriptsBy etc.

CHRLOC:

Chromosome and starting base of associated gene (deprecated for Bioc > 3.1) For this information you should look at a TxDb or OrganismDb object and search for an appropriate field like TXSTART etc. or even better use the associated range based accessors like transcripts or transcriptsBy to get back GRanges objects.

CHRLOCEND:

Chromosome and ending base of associated gene (deprecated for Bioc > 3.1) For this information you should look at a TxDb or OrganismDb object and search for an appropriate field like TXEND etc. or even better use the associated range based accessors like transcripts or transcriptsBy to get back GRanges objects.

COMMON:

Common name

DESCRIPTION:

The description of the associated gene

ENSEMBL:

The ensembl ID as indicated by ensembl

ENSEMBLPROT:

The ensembl protein ID as indicated by ensembl

ENSEMBLTRANS:

The ensembl transcript ID as indicated by ensembl

ENTREZID:

Entrez gene Identifiers

ENZYME:

Enzyme Commission numbers

EVIDENCE:

Evidence codes for GO associations with a gene of interest

EVIDENCEALL:

Evidence codes for GO (includes less specific terms)

GENENAME:

The full gene name

GO:

GO Identifiers associated with a gene of interest

GOALL:

GO Identifiers (includes less specific terms)

INTERPRO:

InterPro identifiers

IPI:

IPI accession numbers

MAP:

cytoband locations

OMIM:

Online Mendelian Inheritance in Man identifiers

ONTOLOGY:

For GO Identifiers, which Gene Ontology (BP, CC, or MF)

ONTOLOGYALL:

Which Gene Ontology (BP, CC, or MF), (includes less specific terms)

ORF:

Yeast ORF Identifiers

PATH:

KEGG Pathway Identifiers

PFAM:

PFAM Identifiers

PMID:

Pubmed Identifiers

PROBEID:

Probe or manufacturer Identifiers for a chip package

PROSITE:

Prosite Identifiers

REFSEQ:

Refseq Identifiers

SGD:

Saccharomyces Genome Database Identifiers

SMART:

Smart Identifiers

SYMBOL:

The official gene symbol

TAIR:

TAIR Identifiers

UNIGENE:

Unigene Identifiers

UNIPROT:

Uniprot Identifiers

To get the latest information about the date stamps and source URLS for the data used to make an annotation package, please use the metadata method as shown in the example below.

Unless otherwise indicated above, the majority of the data for any one package is taken from the source indicated by either it's name (if it's an org package) OR from the name of it's associated org package. So for example, org.Hs.eg.db is using "eg" in the name to indicate that most of the data in that package comes from NCBI entrez gene based data. And org.At.tair.db uses data that primarily comes from tair. For chip packages, the relevant information is the organism package that they depend on. So for example, hgu95av2.db depends on org.Hs.eg.db, and is thus primarily based on NCBI entrez gene ID information.

Author(s)

Marc Carlson

Examples


  library(hgu95av2.db)
  ## List the possible values for columns
  columns(hgu95av2.db)
  ## List the possible values for keytypes
  keytypes(hgu95av2.db)
  ## get some values back
  keys <- head(keys(hgu95av2.db))
  keys
  select(hgu95av2.db, keys=keys, columns=c("SYMBOL","PFAM"),
  keytype="PROBEID")

  ## More infomation about the dates and original sources for these data:
  metadata(hgu95av2.db)

[Package AnnotationDbi version 1.44.0 Index]