makeGOGraph {AnnotationDbi}	R Documentation

A convenience function to generate graphs based on the GO.db package

Description

makeGOGraph is a function to quickly convert any of the three Gene Ontologies in GO.db into a graphNEL object where each edge is given a weight of 1.

Usage


  makeGOGraph(ont = c("bp","mf","cc"))

Arguments

ont
Specifies the ontology: "cc", "bp" or "mf".

Author(s)

Marc Carlson

See Also

GOTerms

Examples


  ## makes a GO graph from the CC ontology 
  f <- makeGOGraph("cc")

[Package AnnotationDbi version 1.44.0 Index]