orgPackageName {AnnotationDbi}	R Documentation

Org package contained in annotation object

Description

Get the name of the org package used by an annotation resource object.

NOTE: This man page is for the orgPackageName S4 generic function defined in the AnnotationDbi package. Bioconductor packages can define specific methods for annotation objects not supported by the default method.

Usage


orgPackageName(x, ...)

Arguments

x
An annotation resource object.

...
Additional arguments.

Value

A character(1) vector indicating the org package name.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

[Package AnnotationDbi version 1.44.0 Index]