print.probetable {AnnotationDbi}	R Documentation

Print method for probetable objects

Description

Prints class(x), nrow(x) and ncol(x), but not the elements of x. The motivation for having this method is that methods from the package base such as print.data.frame will try to print the values of all elements of x, which can take inconveniently much time and screen space if x is large.

Usage


## S3 method for class 'probetable'
print(x, maxrows, ...)

Arguments

x
an object of S3-class probetable.

maxrows
maximum number of rows to print.

...
further arguments that get ignored.

See Also

print.data.frame

Examples


  a = as.data.frame(matrix(runif(1e6), ncol=1e3))
  class(a) = c("probetable", class(a))
  print(a)
  print(as.matrix(a[2:3, 4:6]))

[Package AnnotationDbi version 1.44.0 Index]