toggleProbes {AnnotationDbi}	R Documentation

Methods for getting/setting the filters on a Bimap object

Description

These methods are part of the Bimap interface (see ?Bimap for a quick overview of the Bimap objects and their interface).

Some of these methods are for getting or setting the filtering status on a Bimap object so that the mapping object can toggle between displaying all probes, only single probes (the defualt) or only multiply matching probes.

Other methods are for viewing or setting the filter threshold value on a InpAnnDbBimap object.

Usage

  ## Making a Bimap object that does not prefilter to remove probes that
  ## match multiple genes:
  toggleProbes(x, value)
  hasMultiProbes(x) ##T/F test for exposure of single probes
  hasSingleProbes(x) ##T/F test for exposure of mulitply matched probes

  ## Looking at the SQL filter values for a Bimap
  getBimapFilters(x)

Arguments

x
A Bimap object.

value
A character vector containing the new value that the Bimap should use as the filter. Or the value to toggle a probe mapping to: "all", "single", or "multiple".

Details

toggleProbes(x) is a methods for creating Bimaps that have an alternate filter for which probes get exposed based upon whether these probes map to multiple genes or not.

hasMultiProbes(x) and hasSingleProbes(x) are provided to give a quick test about whether or not such probes are exposed in a given mapping.

getBimapFilters(x) will list all the SQL filters applied to a Bimap object.

Value

A Bimap object of the same subtype as x for exposeAllProbes(x), maskMultiProbes(x) and maskSingleProbes(x).

A TRUE or FALSE value in the case of hasMultiProbes(x) and hasSingleProbes(x).

Author(s)

M. Carlson

See Also

Bimap, Bimap-keys, Bimap-direction, BimapFormatting, Bimap-envirAPI, nhit

Examples

  ## Make a Bimap that contains all the probes
  require("hgu95av2.db")
  mapWithMultiProbes <- toggleProbes(hgu95av2ENTREZID, "all")
  count.mappedLkeys(hgu95av2ENTREZID)
  count.mappedLkeys(mapWithMultiProbes)

  ## Check that it has both multiply and singly matching probes:
  hasMultiProbes(mapWithMultiProbes)
  hasSingleProbes(mapWithMultiProbes)

  ## Make it have Multi probes ONLY:
  OnlyMultiProbes = toggleProbes(mapWithMultiProbes, "multiple")
  hasMultiProbes(OnlyMultiProbes)
  hasSingleProbes(OnlyMultiProbes)

  ## Convert back to a default map with only single probes exposed
  OnlySingleProbes = toggleProbes(OnlyMultiProbes, "single")
  hasMultiProbes(OnlySingleProbes)
  hasSingleProbes(OnlySingleProbes)

[Package AnnotationDbi version 1.56.2 Index]